Moved OverlayElements to mixing
