+++ albumentations/augmentations/functional.py
@@ -1539,19 +1539,3 @@ def generate_approx_gaussian_noise(
 @clipped
 def add_noise(img: np.ndarray, noise: np.ndarray) -> np.ndarray:
     return add_array(img, noise)
-
-
-def copy_and_paste_blend(
-    base_image: np.ndarray,
-    overlay_image: np.ndarray,
-    mask: np.ndarray,
-    offset: Tuple[int, int],
-) -> np.ndarray:
-    y_offset, x_offset = offset
-    blended_image = base_image.copy()
-    mask_indices = np.where(mask > 0)
-    blended_image[mask_indices[0] + y_offset, mask_indices[1] + x_offset] = overlay_image[
-        mask_indices[0],
-        mask_indices[1],
-    ]
-    return blended_image
