+++ albumentations/augmentations/mixing/functional.py
@@ -1,0 +1,19 @@
+from typing import Tuple
+
+import numpy as np
+
+
+def copy_and_paste_blend(
+    base_image: np.ndarray,
+    overlay_image: np.ndarray,
+    mask: np.ndarray,
+    offset: Tuple[int, int],
+) -> np.ndarray:
+    y_offset, x_offset = offset
+    blended_image = base_image.copy()
+    mask_indices = np.where(mask > 0)
+    blended_image[mask_indices[0] + y_offset, mask_indices[1] + x_offset] = overlay_image[
+        mask_indices[0],
+        mask_indices[1],
+    ]
+    return blended_image
