+++ tests/test_functional.py
@@ -1099,29 +1099,3 @@ def test_planckian_jitter_cied():
 
     cied_plankian_jitter = F.planckian_jitter(img, temperature=4500, mode="cied")
     assert np.allclose(cied_plankian_jitter, expected_cied_plankian_jitter, atol=1e-4)
-
-
-@pytest.mark.parametrize("base_image, overlay_image, mask, offset, expected_shape, expected_comparison", [
-    (
-        np.ones((200, 200, 3), dtype=np.uint8) * 255,
-        np.zeros((100, 100, 3), dtype=np.uint8),
-        np.ones((100, 100), dtype=np.uint8) * 255,
-        (50, 50),
-        (200, 200, 3),
-        lambda result, base_image, overlay_image, mask: np.array_equal(result[50:150, 50:150][mask > 0], overlay_image[mask > 0])
-    ),
-    (
-        np.ones((200, 200, 3), dtype=np.uint8) * 255,
-        np.zeros((100, 100, 3), dtype=np.uint8),
-        None,
-        (50, 50),
-        (200, 200, 3),
-        lambda result, base_image, overlay_image, _: np.all(result[50:150, 50:150] != base_image[50:150, 50:150])
-    ),
-])
-def test_copy_and_paste_blend(base_image, overlay_image, mask, offset, expected_shape, expected_comparison):
-    if mask is None:
-        mask = np.ones_like(overlay_image[:, :, 0])
-    result = F.copy_and_paste_blend(base_image, overlay_image, mask, offset)
-    assert result.shape == expected_shape
-    assert expected_comparison(result, base_image, overlay_image, mask)
